Student course registration functionality added.

+++ src/main/java/com/miuapps/miuschedule/service/impl/StudentServiceImpl.java
@@ -1,6 +1,8 @@
 package com.miuapps.miuschedule.service.impl;
 
+import com.miuapps.miuschedule.exceptions.CourseRegisterException;
 import com.miuapps.miuschedule.model.Course;
+import com.miuapps.miuschedule.model.ERole;
 import com.miuapps.miuschedule.model.User;
 import com.miuapps.miuschedule.repository.CourseRepository;
 import com.miuapps.miuschedule.repository.UserRepository;
@@ -31,22 +33,22 @@ public class StudentServiceImpl implements IStudentService {
     }
 
     @Override
-    public void addUser(String userID, String courseID) {
+    public void registerForCourse(String userID, String courseID) throws CourseRegisterException{
         course = courseRepository.findCourseById(courseID);
+        if(course == null) {
+            throw new CourseRegisterException("Given course id does not belongs to a course.");
+        }
         user = userRepository.findUserById(userID);
-        if (course.getUserList().size() < course.getCapacity()) {
+        boolean isStudent = user.getRoles().stream().anyMatch(role -> role.getName().equals(ERole.ROLE_STUDENT));
+        if(!isStudent) {
+            throw new CourseRegisterException("Given user id does not belongs to a student.");
+        }
+        if (course.getUserList() != null && course.getUserList().size() < course.getCapacity()) {
             courseRepository.deleteById(course.getId());
             course.addUser(user);
             courseRepository.save(course);
-        }
-//            if(user.getRoles().equals(ERole.ROLE_STUDENT)){
-//                course.addUser(user);
-//            }else{
-//                //Mesut please implement
-//            }
-//        }
-        else {
-            System.out.println("full capacity");
+        } else {
+            throw new CourseRegisterException("Course capacity is not enough.");
         }
     }
 }
